ai_module/requirements.txt
fastapi
uvicorn
requests
tensorflow
pillow
numpy
python-multipart
scikit-learn
pandas
